moved complete test suite test selection to own module
